+++ BitFaster.Caching/Lru/Builder/AsyncAtomicLruBuilder.cs
@@ -7,11 +7,11 @@ using BitFaster.Caching.Synchronized;
 
 namespace BitFaster.Caching.Lru.Builder
 {
-    public class AtomicFactoryAsyncLruBuilder<K, V> : LruBuilderBase<K, V, AtomicFactoryAsyncLruBuilder<K, V>, IAsyncCache<K, V>>
+    public class AsyncAtomicLruBuilder<K, V> : LruBuilderBase<K, V, AsyncAtomicLruBuilder<K, V>, IAsyncCache<K, V>>
     {
         private readonly ConcurrentLruBuilder<K, AsyncAtomicFactory<K, V>> inner;
 
-        internal AtomicFactoryAsyncLruBuilder(ConcurrentLruBuilder<K, AsyncAtomicFactory<K, V>> inner)
+        internal AsyncAtomicLruBuilder(ConcurrentLruBuilder<K, AsyncAtomicFactory<K, V>> inner)
             : base(inner.info)
         {
             this.inner = inner;
@@ -19,9 +19,8 @@ namespace BitFaster.Caching.Lru.Builder
 
         public override IAsyncCache<K, V> Build()
         {
-            var innerCache = inner.Build();
-
-            return new AtomicFactoryAsyncCache<K, V>(innerCache);
+            var level1 = inner.Build();
+            return new AtomicFactoryAsyncCache<K, V>(level1);
         }
     }
 }
