+++ BitFaster.Caching/Lru/Builder/ScopedAsyncLruBuilder.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BitFaster.Caching.Lru.Builder
+{
+    public sealed class ScopedAsyncLruBuilder<K, V> : LruBuilderBase<K, V, ScopedAsyncLruBuilder<K, V>, IScopedAsyncCache<K, V>> where V : IDisposable
+    {
+        private readonly AsyncConcurrentLruBuilder<K, Scoped<V>> inner;
+
+        internal ScopedAsyncLruBuilder(AsyncConcurrentLruBuilder<K, Scoped<V>> inner)
+            : base(inner.info)
+        {
+            this.inner = inner;
+        }
+
+        ///<inheritdoc/>
+        public override IScopedAsyncCache<K, V> Build()
+        {
+            // this is a legal type conversion due to the generic constraint on W
+            var scopedInnerCache = inner.Build() as IAsyncCache<K, Scoped<V>>;
+
+            return new ScopedAsyncCache<K, V>(scopedInnerCache);
+        }
+    }
+}
