+++ BitFaster.Caching/Lru/Builder/ScopedLruBuilder.cs
@@ -26,44 +26,4 @@ namespace BitFaster.Caching.Lru.Builder
             return new ScopedCache<K, V>(scopedInnerCache);
         }
     }
-
-    public sealed class ScopedAsyncLruBuilder<K, V> : LruBuilderBase<K, V, ScopedAsyncLruBuilder<K, V>, IScopedAsyncCache<K, V>> where V : IDisposable
-    {
-        private readonly AsyncConcurrentLruBuilder<K, Scoped<V>> inner;
-
-        internal ScopedAsyncLruBuilder(AsyncConcurrentLruBuilder<K, Scoped<V>> inner)
-            : base(inner.info)
-        {
-            this.inner = inner;
-        }
-
-        ///<inheritdoc/>
-        public override IScopedAsyncCache<K, V> Build()
-        {
-            // this is a legal type conversion due to the generic constraint on W
-            var scopedInnerCache = inner.Build() as IAsyncCache<K, Scoped<V>>;
-
-            return new ScopedAsyncCache<K, V>(scopedInnerCache);
-        }
-    }
-
-    public sealed class ScopedAsyncAtomicLruBuilder<K, V> : LruBuilderBase<K, V, ScopedAsyncAtomicLruBuilder<K, V>, IScopedAsyncCache<K, V>> where V : IDisposable
-    {
-        private readonly AsyncConcurrentLruBuilder<K, ScopedAsyncAtomicFactory<K, V>> inner;
-
-        internal ScopedAsyncAtomicLruBuilder(AsyncConcurrentLruBuilder<K, ScopedAsyncAtomicFactory<K, V>> inner)
-            : base(inner.info)
-        {
-            this.inner = inner;
-        }
-
-        ///<inheritdoc/>
-        public override IScopedAsyncCache<K, V> Build()
-        {
-            // this is a legal type conversion due to the generic constraint on W
-            var scopedInnerCache = inner.Build() as ICache<K, ScopedAsyncAtomicFactory<K, V>>;
-
-            return new AtomicFactoryScopedAsyncCache<K, V>(scopedInnerCache);
-        }
-    }
 }
